Update fmrib_fastr.m
Handle edge case when there isn't enough recording after the last TR trigger in a data section. The old msg saying "If you get an error "line 746 of fmrib_fastr: index exceeds matrix dimensions" it means there is an inconsistency in your TR triggers, either the TR length or the number of markers" is not informative since the error is silent and occurs even when a recording has consistent TR intervals.

Added more informative error handling